+++ frontend/src/App.js
@@ -1,16 +1,13 @@
 import './css/App.css';
-import './css/styles.css';
-import { BrowserRouter as Router, Route, Switch, Link, BrowserRouter, Routes } from 'react-router-dom';
-import './webroute/Home';
+// import './css/styles.css';
 import './webroute/Ticket';
-import Home from './webroute/Home';
 import Ticket from './webroute/Ticket';
-import TicketQueue from './webroute/TicketQueue';
 import TicketInfo from './webroute/TicketInfo';
-import { useState } from 'react';
 import { Routes as ReactRoutes, Route as ReactRoute} from 'react-router-dom';
 import Login from './webroute/Login';
 import Admin from './webroute/Admin';
+import UserProfile from './webroute/userProfile';
+import Home from './webroute/home'
 
 function App() {
 
@@ -20,10 +17,10 @@ function App() {
       <ReactRoutes>
         <ReactRoute path='/' element={<Login/>}/>
         <ReactRoute path='/Ticket' element={<Ticket/>}/>
-        <ReactRoute path='/TicketQueue' element={<TicketQueue/>} />
+        <ReactRoute path='/home' element={<Home/>} />
         <ReactRoute path='/TicketInfo' element={<TicketInfo/>}  />
         <ReactRoute path='/Admin' element={<Admin/>}/>
-        
+        <ReactRoute path='/profile' element={<UserProfile/>}/>
       </ReactRoutes>
     </div>
   );
