+++ frontend/src/components/footer.js
@@ -1,0 +1,16 @@
+import {React} from 'react'
+
+const Footer = ()=>{
+    return(
+    <>
+    <footer>
+        this is a footer
+    </footer>
+    </>
+    );
+}
+
+
+
+
+export default Footer
