+++ frontend/src/components/header.js
@@ -1,0 +1,16 @@
+import {React} from 'react'
+
+const Header = ()=>{
+    return(
+    <>
+    <header>
+        this is a header
+    </header>
+    </>
+    )
+}
+
+
+
+
+export default Header
