+++ frontend/src/components/temComponent.js
@@ -1,0 +1,14 @@
+import React from "react";
+import '../css/temComponent.css'
+
+const TempComponent = ()=>{
+    return(
+        <div className="TempComponent">
+          <div className="box">
+            TempComponent
+          </div>
+        </div>
+    )
+}
+
+export default TempComponent
