+++ frontend/src/css/App.css
@@ -1,6 +1,4 @@
-.App {
-  /* text-align: center; */
-}
+
 
 .App-logo {
   height: 40vmin;
