+++ frontend/src/css/Login.css
@@ -1,60 +1,69 @@
+body, html {
+  height: 100%;
+  margin: 0;
+}
+
 .container {
-    background-color: #fff;
-    border-radius: 10px;
-    box-shadow: 0 0 20px rgba(0, 0, 0, 0.1);
-    padding: 30px;
-    max-width: 400px;
-    width: 100%;
-  }
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  background-color: #fff;
+  border-radius: 10px;
+  box-shadow: 0 0 20px rgba(0, 0, 0, 0.1);
+  padding: 30px;
+  max-width: 400px;
+  width: calc(100% - 40px);
+}
 
-  h2 {
-    margin-top: 0;
-    margin-bottom: 20px;
-    text-align: center;
-  }
+h2 {
+  margin-top: 0;
+  margin-bottom: 20px;
+  text-align: center;
+}
 
-  label {
-    display: block;
-    margin-bottom: 10px;
-    font-weight: bold;
-  }
+label {
+  display: block;
+  margin-bottom: 10px;
+  font-weight: bold;
+}
 
-  input[type="email"],
-  input[type="password"] {
-    width: 100%;
-    padding: 10px;
-    margin-bottom: 20px;
-    border-radius: 5px;
-    border: 1px solid #ccc;
-    box-sizing: border-box;
-  }
+input[type="email"],
+input[type="password"],
+select {
+  width: 100%;
+  padding: 10px;
+  margin-bottom: 20px;
+  border-radius: 5px;
+  border: 1px solid #ccc;
+  box-sizing: border-box;
+}
 
-  input[type="submit"] {
-    background-color: #007bff;
-    border: none;
-    color: #fff;
-    padding: 12px 20px;
-    border-radius: 5px;
-    cursor: pointer;
-    font-size: 16px;
-    width: 100%;
-    display: block;
-  }
+input[type="submit"] {
+  background-color: #007bff;
+  border: none;
+  color: #fff;
+  padding: 12px 20px;
+  border-radius: 5px;
+  cursor: pointer;
+  font-size: 16px;
+  width: 100%;
+}
 
-  input[type="submit"]:hover {
-    background-color: #0056b3;
-  }
+input[type="submit"]:hover {
+  background-color: #0056b3;
+}
 
-  .form-footer {
-    text-align: center;
-    margin-top: 20px;
-  }
+.form-footer {
+  text-align: center;
+  margin-top: 20px;
+}
 
-  .form-footer a {
-    color: #007bff;
-    text-decoration: none;
-  }
+.form-footer a {
+  color: #007bff;
+  text-decoration: none;
+}
 
-  .form-footer a:hover {
-    text-decoration: underline;
-  }
+.form-footer a:hover {
+  text-decoration: underline;
+}
