+++ frontend/src/css/SideNav.css
@@ -1,13 +1,4 @@
-.sidenav {
-    height: 100%;
-    width: 220px; /* Adjust width as needed */
-    position: fixed;
-    top: 0;
-    left: 0;
-    background-color: black; /* Background color */
-    overflow-x: hidden;
-    padding-top: 20px; /* Space at the top */
-  }
+
   
   .sidenav a {
     padding: 10px 15px; /* Padding for links */
@@ -23,6 +14,7 @@
   a{
     color: white;
     font-weight: bold;
+    cursor: pointer;
   }
   
   .user{
@@ -31,11 +23,13 @@
   }
   .user_icon{
     width: 80%;
-    height: 100px;
+    height: 80px;
+    margin-bottom: 20px;
     display: flex;
     align-items: center;
     margin-left: 1rem;
     font-size: 13px;
+    cursor: pointer;
   }
   .barane{
     display: flex;
@@ -43,3 +37,63 @@
     align-items: center;
     margin-top: 12px;
   }
+
+.logout-box::backdrop{
+  background-color: rgba(0, 0, 0, 0.5);
+}
+
+.logout-box {
+  width: 330px;
+  padding: 20px;
+  background-color: #ffffff;
+  border: 1px solid #ccc;
+  border-radius: 13px;
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.6);
+  text-align: center;
+}
+
+/* .logout-box h2 {
+  font-size: 19px;
+  color: #333333;
+} */
+
+.logout-box .confirm {
+  text-align: left;
+  font-size: 25px;
+  font-weight: 500;
+}
+
+.logout-box #confirmButtons {
+  display: flex;
+  justify-content: space-between;
+  margin-top: 20px;
+}
+
+.logout-box button {
+  border-radius: 13px;
+  width: 40%;
+  padding: 10px 20px;
+  font-size: 16px;
+  color: #ffffff;
+  border: 1px solid #ccc;
+  border-radius: 5px;
+  cursor: pointer;
+}
+
+.logout-box #yes{
+  background-color: #0056b3;
+}
+
+.logout-box #no{
+  color: black;
+}
+
+.logout-box #yes:hover {
+  background-color: #3681d1;
+}
+
+.logout-box #no:hover {
+  background-color: #0056b3;
+  color: white;
+}
+
