+++ frontend/src/css/home.css
@@ -1,0 +1,20 @@
+.home {
+    display: flex;
+  }
+  
+  .sidenav {
+    width: 16%;
+    position: fixed;
+    top: 0;
+    left: 0;
+    bottom: 0;
+    background-color: black;
+  }
+  
+  .componentContainer {
+    flex: 1;
+    margin-left: 16%;
+    padding: 20px;
+    overflow-y: auto;
+  }
+
