+++ frontend/src/webroute/Admin.js
@@ -1,24 +1,25 @@
 import React from "react";
-import useNavigate from 'react'
+import Footer from '../components/footer';
+import Header from '../components/header';
 const Admin = () => {
 
-    const navigate = useNavigate();
-    const user = JSON.parse(localStorage.getItem("user"));
-    if (user === null){
-      window.location.replace("/");
-    }
+    // const navigate = useNavigate();
+    // const user = JSON.parse(localStorage.getItem("user"));
+    // if (user === null){
+    //   window.location.replace("/");
+    // }
     
-    // Example usage:
-    const userCookie = JSON.parse(localStorage.getItem('user'));
-    if (userCookie) {
-        console.log('User cookie:', userCookie);
-    } else {
-        console.log('User cookie not found');
-    }
+    // // Example usage:
+    // const userCookie = JSON.parse(localStorage.getItem('user'));
+    // if (userCookie) {
+    //     console.log('User cookie:', userCookie);
+    // } else {
+    //     console.log('User cookie not found');
+    // }
 
     return (
         <div>Hello World
-             {userCookie.id}
+             {/* {userCookie.id} */}
         </div>
     );
 }
